breakup settings and use tabs
