+++ frontend/app/settings/_components/anthropic-settings-dialog.tsx
@@ -92,7 +92,7 @@ const AnthropicSettingsDialog = ({
         action: {
           label: "Settings",
           onClick: () => {
-            router.push("/settings?focusLlmModel=true");
+            router.push("/settings/langflow?focusLlmModel=true");
           },
         },
       });
