+++ frontend/app/settings/_components/ollama-settings-dialog.tsx
@@ -102,7 +102,7 @@ const OllamaSettingsDialog = ({
         action: {
           label: "Settings",
           onClick: () => {
-            router.push("/settings?focusLlmModel=true");
+            router.push("/settings/langflow?focusLlmModel=true");
           },
         },
       });
