+++ frontend/app/settings/_components/watsonx-settings-dialog.tsx
@@ -105,7 +105,7 @@ const WatsonxSettingsDialog = ({
         action: {
           label: "Settings",
           onClick: () => {
-            router.push("/settings?focusLlmModel=true");
+            router.push("/settings/langflow?focusLlmModel=true");
           },
         },
       });
