+++ frontend/app/settings/layout.tsx
@@ -1,0 +1,14 @@
+import { Suspense } from "react";
+import { SettingsShell } from "./_components/settings-shell";
+
+export default function SettingsLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <Suspense>
+      <SettingsShell>{children}</SettingsShell>
+    </Suspense>
+  );
+}
