+++ frontend/app/settings/page.tsx
@@ -1,37 +1,5 @@
-import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
-import { getQueryClient } from "@/app/api/get-query-client";
-import { fetchFromBackend } from "@/lib/fetch-server";
-import SettingsClient from "./settings-client";
+import { redirect } from "next/navigation";
 
-export default async function SettingsPage() {
-  const queryClient = getQueryClient();
-
-  try {
-    await Promise.all([
-      queryClient.prefetchQuery({
-        queryKey: ["settings"],
-        queryFn: async () => {
-          const res = await fetchFromBackend("settings");
-          if (!res.ok) throw new Error("Failed to fetch settings");
-          return res.json();
-        },
-      }),
-      queryClient.prefetchQuery({
-        queryKey: ["api-keys"],
-        queryFn: async () => {
-          const res = await fetchFromBackend("keys");
-          if (!res.ok) throw new Error("Failed to fetch api keys");
-          return res.json();
-        },
-      }),
-    ]);
-  } catch {
-    // Backend unavailable or unauthenticated — client handles loading/auth normally
-  }
-
-  return (
-    <HydrationBoundary state={dehydrate(queryClient)}>
-      <SettingsClient />
-    </HydrationBoundary>
-  );
+export default function SettingsPage() {
+  redirect("/settings/connectors");
 }
