+++ frontend/package.json
@@ -45,6 +45,7 @@
     "motion": "^12.23.12",
     "next": "^16.1.7",
     "next-themes": "^0.4.6",
+    "radix-ui": "^1.4.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-dropzone": "^14.3.8",
